requirements.txt
numpy>=1.19.0
PyGithub>=1.55
pyyaml>=5.4
pathlib2>=2.3.0
requests>=2.25.0
